Initial commit of a working implementation of the patrol algorithm
Problem:

This project is an implementation of the Skeleton-based multi-agent opponent search
algorithm whitepaper. This can be iterated on separately from other game projects
as a sandbox to test new ideas for AI, or improvements on the patrol algorithm itself.

Solution:

Implementation was done using some reusable BCI plugins and following the paper.

Changes:

* Adding all BCI plugins
* Adding the Writerside documentation
* Adding the implementation

Testing:

Tested manually in the editor.