Adding minimal project content
Problem:

The project currently cannot be checked out from GitHub due to no
content. Until content is checked in, the project cannot be used
by other people.

Solution:

This change adds the minimally used set of content from the
start project, as well as the custom blueprints. A signficant
amount of starter content was cut so as to keep the size of the
repo under 1gb. Considering it is highly likely no textures, models
or any other heavy content will be created for this project,
this should be reasonable for development in the long term.

Changes:

* Checked in all used starter content
* Checked in custom blueprints

Testing:

Tested that the project can still open and run.